+++ src/Propulsion.EventStore/EventStoreSink.fs
@@ -21,81 +21,76 @@ module Internal =
         type [<NoComparison; NoEquality; RequireQualifiedAccess>] Result =
             | Ok of updatedPos: int64
             | Duplicate of updatedPos: int64
-            | PartialDuplicate of overage: Event[]
-            | PrefixMissing of batch: Event[] * writePos: int64
+            | PartialDuplicate of updatedPos: int64
+            | PrefixMissing of gap: int * actualPos: int64
 
-        let logTo (log: ILogger) (res: FsCodec.StreamName * Result<struct (StreamSpan.Metrics * Result), struct (StreamSpan.Metrics * exn)>) =
-            match res with
-            | stream, Ok (_, Result.Ok pos) ->
+        let logTo (log: ILogger) (stream: FsCodec.StreamName): Result<Dispatcher.ResProgressAndMetrics<Result>, Dispatcher.ExnAndMetrics> -> unit = function
+            | Ok (Result.Ok pos, _, _) ->
                 log.Information("Wrote     {stream} up to {pos}", stream, pos)
-            | stream, Ok (_, Result.Duplicate updatedPos) ->
+            | Ok (Result.Duplicate updatedPos, _, _) ->
                 log.Information("Ignored   {stream} (synced up to {pos})", stream, updatedPos)
-            | stream, Ok (_, Result.PartialDuplicate overage) ->
-                log.Information("Requeuing {stream} {pos} ({count} events)", stream, overage[0].Index, overage.Length)
-            | stream, Ok (_, Result.PrefixMissing (batch, pos)) ->
-                log.Information("Waiting   {stream} missing {gap} events ({count} events @ {pos})",
-                                stream, batch[0].Index - pos, batch.Length, batch[0].Index)
-            | stream, Error (_, exn) ->
+            | Ok (Result.PartialDuplicate updatedPos, _, _) ->
+                log.Information("Requeuing {stream} {pos}", stream, updatedPos)
+            | Ok (Result.PrefixMissing (gap, pos), _, _) ->
+                log.Information("Waiting   {stream} missing {gap} events before {pos}", stream, gap, pos)
+            | Error (exn, _, _) ->
                 log.Warning(exn,"Writing   {stream} failed, retrying", stream)
 
         let write (log: ILogger) (context: EventStoreContext) stream (span: Event[]) ct = task {
-            log.Debug("Writing {s}@{i}x{n}", stream, span[0].Index, span.Length)
+            let i = StreamSpan.index span
+            log.Debug("Writing {s}@{i}x{n}", stream, i, span.Length)
 #if EVENTSTORE_LEGACY
-            let! res = context.Sync(log, stream, span[0].Index - 1L, span |> Array.map (fun span -> span :> _))
+            let! res = context.Sync(log, stream, i - 1L, span |> Array.map (fun span -> span :> _))
 #else
-            let! res = context.Sync(log, stream, span[0].Index - 1L, span |> Array.map (fun span -> span :> _), ct)
+            let! res = context.Sync(log, stream, i - 1L, span |> Array.map (fun span -> span :> _), ct)
 #endif
-            let ress =
+            let res' =
                 match res with
                 | GatewaySyncResult.Written (Token.Unpack pos') ->
                     Result.Ok (pos'.streamVersion + 1L)
                 | GatewaySyncResult.ConflictUnknown (Token.Unpack pos) ->
                     match pos.streamVersion + 1L with
-                    | actual when actual < span[0].Index -> Result.PrefixMissing (span, actual)
-                    | actual when actual >= span[0].Index + span.LongLength -> Result.Duplicate actual
-                    | actual -> Result.PartialDuplicate (span |> Array.skip (actual - span[0].Index |> int))
-            log.Debug("Result: {res}", ress)
-            return ress }
+                    | actual when actual < i -> Result.PrefixMissing (i - actual |> int, actual)
+                    | actual when actual >= i + span.LongLength -> Result.Duplicate actual
+                    | actual -> Result.PartialDuplicate actual
+            log.Debug("Result: {res}", res')
+            return res' }
 
         type [<RequireQualifiedAccess>] ResultKind = TimedOut | Other
 
     type Dispatcher =
 
-        static member Create(log: ILogger, storeLog, connections: _[], maxDop) =
-            let writerResultLog = log.ForContext<Writer.Result>()
+        static member Create(storeLog, connections: _[], maxDop) =
             let mutable robin = 0
 
-            let attemptWrite stream span ct = task {
+            let attemptWrite stream span _revision ct = task {
                 let index = System.Threading.Interlocked.Increment(&robin) % connections.Length
                 let selectedConnection = connections[index]
                 let maxEvents, maxBytes = 65536, 4 * 1024 * 1024 - (*fudge*)4096
-                let struct (met, span') = StreamSpan.slice Event.renderedSize (maxEvents, maxBytes) span
-                try let! res = Writer.write storeLog selectedConnection (FsCodec.StreamName.toString stream) span' ct
-                    return Ok struct (met, res)
-                with e -> return Error struct (met, e) }
-            let interpretProgress (streams: Scheduling.StreamStates<_>) stream res =
-                let applyResultToStreamState = function
-                    | Ok struct (_stats, Writer.Result.Ok pos) ->               streams.RecordWriteProgress(stream, pos, null)
-                    | Ok (_stats, Writer.Result.Duplicate pos) ->               streams.RecordWriteProgress(stream, pos, null)
-                    | Ok (_stats, Writer.Result.PartialDuplicate overage) ->    streams.RecordWriteProgress(stream, overage[0].Index, [| overage |])
-                    | Ok (_stats, Writer.Result.PrefixMissing (overage, pos)) -> streams.RecordWriteProgress(stream, pos, [| overage |])
-                    | Error struct (_stats, _exn) ->                            streams.SetMalformed(stream, false)
-                let ss = applyResultToStreamState res
-                Writer.logTo writerResultLog (stream, res)
-                struct (ss.WritePos, res)
-            Dispatcher.Concurrent<_, _, _, _>.Create(maxDop, attemptWrite, interpretProgress)
+                let struct (span, met) = StreamSpan.slice Event.renderedSize (maxEvents, maxBytes) span
+                try let! res = Writer.write storeLog selectedConnection (FsCodec.StreamName.toString stream) span ct
+                    let hp = res |> function
+                        | Writer.Result.Ok pos' | Writer.Result.Duplicate pos' | Writer.Result.PartialDuplicate pos' -> Buffer.HandlerProgress.ofPos pos' |> ValueSome
+                        | Writer.Result.PrefixMissing _ -> ValueNone
+                    return Ok struct (res, hp, met)
+                with e -> return Error struct (e, false, met) }
+            let interpretProgress = function
+                | Ok struct (_res, hp, _met) as res -> struct (res, hp, false)
+                | Error struct (_exn, malformed, _met) as res -> res, ValueNone, malformed
+            Dispatcher.Concurrent<_, _, _, _>.Create(maxDop, project = attemptWrite, interpretProgress = interpretProgress)
 
 type WriterResult = Internal.Writer.Result
 
 type EventStoreSinkStats(log: ILogger, statsInterval, stateInterval, [<O; D null>] ?failThreshold, [<O; D null>] ?logExternalStats) =
-    inherit Scheduling.Stats<struct (StreamSpan.Metrics * WriterResult), struct (StreamSpan.Metrics * exn)>(
+    inherit Scheduling.Stats<Dispatcher.ResProgressAndMetrics<WriterResult>, Dispatcher.ExnAndMetrics>(
         log, statsInterval, stateInterval, ?failThreshold = failThreshold,
         logExternalStats = defaultArg logExternalStats Log.InternalMetrics.dump)
+    let writerResultLog = log.ForContext<WriterResult>()
     let mutable okStreams, okEvents, okBytes, exnCats, exnStreams, exnEvents, exnBytes = HashSet(), 0, 0L, Stats.Counters(), HashSet(), 0 , 0L
     let mutable resultOk, resultDup, resultPartialDup, resultPrefix, resultExn = 0, 0, 0, 0, 0
     override _.Handle message =
         match message with
-        | { stream = stream; result = Ok ((es, bs), res) } ->
+        | { stream = stream; result = Ok (res, _hp, (es, us, bs)) as r } ->
             okStreams.Add stream |> ignore
             okEvents <- okEvents + es
             okBytes <- okBytes + int64 bs
@@ -104,13 +99,15 @@ type EventStoreSinkStats(log: ILogger, statsInterval, stateInterval, [<O; D null
             | WriterResult.Duplicate _ -> resultDup <- resultDup + 1
             | WriterResult.PartialDuplicate _ -> resultPartialDup <- resultPartialDup + 1
             | WriterResult.PrefixMissing _ -> resultPrefix <- resultPrefix + 1
-            base.RecordOk(message)
-        | { stream = stream; result = Error ((es, bs), Exception.Inner exn) } ->
+            Internal.Writer.logTo writerResultLog stream r
+            base.RecordOk(message, us <> 0)
+        | { stream = stream; result = Error (Exception.Inner exn, _malformed, (es, _us, bs)) as r } ->
             exnCats.Ingest(StreamName.categorize stream)
             exnStreams.Add stream |> ignore
             exnEvents <- exnEvents + es
             exnBytes <- exnBytes + int64 bs
             resultExn <- resultExn + 1
+            Internal.Writer.logTo writerResultLog stream r
             base.RecordExn(message, OutcomeKind.classify exn, log.ForContext("stream", stream).ForContext("events", es), exn)
     override _.DumpStats() =
         let results = resultOk + resultDup + resultPartialDup + resultPrefix
@@ -138,7 +135,7 @@ type EventStoreSink =
             // Tune the sleep time when there are no items to schedule or responses to process. Default 1ms.
             ?idleDelay,
             ?ingesterStateInterval, ?commitInterval): SinkPipeline =
-        let dispatcher = Internal.Dispatcher.Create(log, storeLog, connections, maxConcurrentStreams)
+        let dispatcher = Internal.Dispatcher.Create(storeLog, connections, maxConcurrentStreams)
         let scheduler =
             let dumpStreams logStreamStates _log = logStreamStates Event.storedSize
             Scheduling.Engine(dispatcher, stats, dumpStreams, pendingBufferSize = 5, ?purgeInterval = purgeInterval, ?idleDelay = idleDelay)