+++ src/Propulsion/Sync.fs
@@ -9,14 +9,14 @@ open System.Collections.Generic
 
 [<AbstractClass>]
 type Stats<'Outcome>(log: ILogger, statsInterval, stateInterval, [<O; D null>] ?failThreshold) =
-    inherit Scheduling.Stats<struct (StreamSpan.Metrics * TimeSpan * 'Outcome), struct (StreamSpan.Metrics * exn)>(log, statsInterval, stateInterval, ?failThreshold = failThreshold)
-    let mutable okStreams, okEvents, okBytes, exnStreams, exnEvents, exnBytes = HashSet(), 0, 0L, HashSet(), 0, 0L
+    inherit Scheduling.Stats<struct ('Outcome * StreamSpan.Metrics * TimeSpan), Dispatcher.ExnAndMetrics>(log, statsInterval, stateInterval, ?failThreshold = failThreshold)
+    let mutable okStreams, okEvents, okUnfolds, okBytes, exnStreams, exnEvents, exnUnfolds, exnBytes = HashSet(), 0, 0, 0L, HashSet(), 0, 0, 0L
     let prepareStats = Stats.LatencyStats("prepare")
     override _.DumpStats() =
         if okStreams.Count <> 0 && exnStreams.Count <> 0 then
-            log.Information("Completed {okMb:n0}MB {okStreams:n0}s {okEvents:n0}e Exceptions {exnMb:n0}MB {exnStreams:n0}s {exnEvents:n0}e",
-                            Log.miB okBytes, okStreams.Count, okEvents, Log.miB exnBytes, exnStreams.Count, exnEvents)
-        okStreams.Clear(); okEvents <- 0; okBytes <- 0L; exnStreams.Clear(); exnBytes <- 0; exnEvents <- 0
+            log.Information("Completed {okMb:n0}MB {okStreams:n0}s {okEvents:n0}e {okUnfolds:n0}u Exceptions {exnMb:n0}MB {exnStreams:n0}s {exnEvents:n0}e {exnUnfolds:n0}u",
+                            Log.miB okBytes, okStreams.Count, okEvents, okUnfolds, Log.miB exnBytes, exnStreams.Count, exnEvents, exnUnfolds)
+        okStreams.Clear(); okEvents <- 0; okUnfolds <- 0; okBytes <- 0L; exnStreams.Clear(); exnBytes <- 0; exnEvents <- 0; exnUnfolds <- 0
         prepareStats.Dump log
 
     abstract member Classify: exn -> OutcomeKind
@@ -24,18 +24,20 @@ type Stats<'Outcome>(log: ILogger, statsInterval, stateInterval, [<O; D null>] ?
 
     override this.Handle message =
         match message with
-        | { stream = stream; result = Ok ((es, bs), prepareElapsed, outcome) } ->
+        | { stream = stream; result = Ok (outcome, (es, us, bs), prepareElapsed) } ->
             okStreams.Add stream |> ignore
             okEvents <- okEvents + es
+            okUnfolds <- okUnfolds + us
             okBytes <- okBytes + int64 bs
             prepareStats.Record prepareElapsed
-            base.RecordOk message
+            base.RecordOk(message, us <> 0)
             this.HandleOk outcome
-        | { stream = stream; result = Error ((es, bs), Exception.Inner exn) } ->
+        | { stream = stream; result = Error (Exception.Inner exn, _malformed, (es, us, bs)) } ->
             exnStreams.Add stream |> ignore
             exnEvents <- exnEvents + es
+            exnUnfolds <- exnUnfolds + us
             exnBytes <- exnBytes + int64 bs
-            base.RecordExn(message, this.Classify exn, log.ForContext("stream", stream).ForContext("events", es), exn)
+            base.RecordExn(message, this.Classify exn, log.ForContext("stream", stream).ForContext("events", es).ForContext("unfolds", us), exn)
 
     abstract member HandleOk: outcome: 'Outcome -> unit
 
@@ -44,34 +46,30 @@ type Factory private () =
 
     static member StartAsync
         (   log: ILogger, maxReadAhead, maxConcurrentStreams,
-            handle: Func<FsCodec.StreamName, FsCodec.ITimelineEvent<'F>[], CancellationToken, Task<struct ('R * 'Outcome)>>,
-            toIndex: Func<FsCodec.ITimelineEvent<'F>[], 'R, int64>,
+            handle: Func<FsCodec.StreamName, FsCodec.ITimelineEvent<'F>[], CancellationToken, Task<struct ('Outcome * int64)>>,
             stats: Stats<'Outcome>, sliceSize, eventSize,
             ?dumpExternalStats, ?idleDelay, ?maxBytes, ?maxEvents, ?purgeInterval, ?ingesterStateInterval, ?commitInterval)
         : SinkPipeline<Ingestion.Ingester<StreamEvent<'F> seq>> =
 
         let maxEvents, maxBytes = defaultArg maxEvents 16384, (defaultArg maxBytes (1024 * 1024 - (*fudge*)4096))
 
-        let attemptWrite stream (events: FsCodec.ITimelineEvent<'F>[]) ct = task {
-            let struct (met, span') = StreamSpan.slice<'F> sliceSize (maxEvents, maxBytes) events
+        let attemptWrite stream (events: FsCodec.ITimelineEvent<'F>[]) revision ct = task {
+            let struct (trimmed, met) = StreamSpan.slice<'F> sliceSize (maxEvents, maxBytes) events
             let prepareTs = Stopwatch.timestamp ()
-            try let! res, outcome = handle.Invoke(stream, span', ct)
-                let index' = toIndex.Invoke(span', res)
-                return Ok struct (index', met, Stopwatch.elapsed prepareTs, outcome)
-            with e -> return Error struct (met, e) }
+            try let! outcome, index' = handle.Invoke(stream, trimmed, ct)
+                return Ok struct (outcome, Buffer.HandlerProgress.ofMetricsAndPos revision met index', met, Stopwatch.elapsed prepareTs)
+            with e -> return Error struct (e, false, met) }
 
-        let interpretProgress _streams (stream: FsCodec.StreamName) = function
-            | Ok struct (i', met, prep, outcome) -> struct (ValueSome i', Ok struct (met, prep, outcome))
-            | Error struct (struct (eventCount, bytesCount) as met, exn: exn) ->
-                log.Warning(exn, "Handling {events:n0}e {bytes:n0}b for {stream} failed, retrying", eventCount, bytesCount, stream)
-                ValueNone, Error struct (met, exn)
+        let interpretProgress = function
+            | Ok struct (outcome, hp, met: StreamSpan.Metrics, prep) -> struct (Ok struct (outcome, met, prep), ValueSome hp, false)
+            | Error struct (exn: exn, malformed, met) -> Error struct (exn, malformed, met), ValueNone, malformed
 
-        let dispatcher: Scheduling.IDispatcher<_, _, _, _> = Dispatcher.Concurrent<_, _, _, _>.Create(maxConcurrentStreams, attemptWrite, interpretProgress)
+        let dispatcher: Scheduling.IDispatcher<_, _, _, _> = Dispatcher.Concurrent<_, _, _, _>.Create(maxConcurrentStreams, attemptWrite, interpretProgress = interpretProgress)
         let dumpStreams logStreamStates log =
             logStreamStates eventSize
             match dumpExternalStats with Some f -> f log | None -> ()
         let scheduler =
-            Scheduling.Engine<struct (int64 * StreamSpan.Metrics * TimeSpan * 'Outcome), struct (StreamSpan.Metrics * TimeSpan * 'Outcome), struct (StreamSpan.Metrics * exn), 'F>
+            Scheduling.Engine<struct ('Outcome * Buffer.HandlerProgress * StreamSpan.Metrics * TimeSpan), struct ('Outcome * StreamSpan.Metrics * TimeSpan), Dispatcher.ExnAndMetrics, 'F>
                 (dispatcher, stats, dumpStreams, pendingBufferSize = maxReadAhead, ?idleDelay = idleDelay, ?purgeInterval = purgeInterval)
 
         Factory.Start(log, scheduler.Pump, maxReadAhead, scheduler,